+++ R/lazyframe__lazy.R
@@ -423,8 +423,8 @@ LazyFrame_collect = function(
 #'
 #' It is useful to not block the R session while query executes. If you use
 #' [`<Expr>$map()`][Expr_map] or [`<Expr>apply()`][Expr_apply] to run R functions
-#' in the query, then you must pass `in_background = TRUE` in `$map()` (or
-#' `$apply()`). Otherwise, `$collect_in_background()` will fail because the main
+#' in the query, then you must pass `in_background = TRUE` in `$map_batches()` (or
+#' `$map_elements()`). Otherwise, `$collect_in_background()` will fail because the main
 #' R session is not available for polars execution. See also examples below.
 #'
 #' @keywords LazyFrame DataFrame_new
@@ -1404,7 +1404,7 @@ LazyFrame_fetch = function(
 #' pl$LazyFrame(iris)$
 #'   sort("Sepal.Length")$
 #'   group_by("Species", maintain_order = TRUE)$
-#'   agg(pl$col(pl$Float64)$apply(r_func))$
+#'   agg(pl$col(pl$Float64)$map_elements(r_func))$
 #'   profile()
 LazyFrame_profile = function(
     type_coercion = TRUE,