+++ R/series__series.R
@@ -350,24 +350,27 @@ Series_value_counts = function(sort = TRUE, parallel = FALSE) {
 #' @return Series
 #' @keywords Series
 #' @aliases apply
-#' @name Series_apply
 #'
 #' @examples
 #' s = pl$Series(letters[1:5], "ltrs")
 #' f = \(x) paste(x, ":", as.integer(charToRaw(x)))
-#' s$apply(f, pl$Utf8)
+#' s$map_elements(f, pl$Utf8)
 #'
 #' # same as
 #' pl$Series(sapply(s$to_r(), f), s$name)
-Series_apply = function(
+Series_map_elements = function(
     fun, datatype = NULL, strict_return_type = TRUE, allow_fail_eval = FALSE) {
-  unwrap(.pr$Series$apply(
+  .pr$Series$map_elements(
     self, fun, datatype, strict_return_type, allow_fail_eval
-  ), "in $apply():")
+  ) |> unwrap("in $map_elements():")
 }
 
-
-
+Series_apply = function(f, return_type = NULL, strict_return_type = TRUE,
+                      allow_fail_eval = FALSE, in_background = FALSE) {
+  warning("$apply() is deprecated and will be removed in 0.12.0. Use $map_elements() instead.")
+  Series_map_elements(f, return_type = return_type, strict_return_type = strict_return_type,
+                    allow_fail_eval = allow_fail_eval, in_background = in_background)
+}
 
 #' Series_len
 #' @description Length of this Series.